fix: centralize second formatter
