+++ project-handbook.md
@@ -87,6 +87,7 @@
 - **2025/11/24**
   - ✅ 完成：新增演出模式告警模組與單元測試，支援 T-30/T-10/通過中階段資訊提示。
   - ✅ 完成：更新 Demo 面板 UI，加入告警橫幅與動態標示，讓第一架航機狀態更一目了然。
+  - ✅ 完成：統一秒數格式化工具，收斂重複實作並補上共用測試。
 - **2025/11/23**
   - ✅ 完成：將噪音等級閾值改為 ObservationSite 可設定參數並加入測試覆蓋，回應 Gemini 的彈性需求。
   - ✅ 完成：重構 pass-processing 為 flatMap 流程以減少暫存陣列，對齊先前的優化計畫。
